+++ docs/ChangeLog.rst
@@ -1,7 +1,20 @@
-.. default-role:: literal
+.. default-role:: py:obj
+.. highlight:: none
+.. py:currentmodule:: factoriocalc
 
 ChangeLog
 =========
+
+Changes from 0.2.1 to 0.3.0 (???)
+------------------------------------------
+
+* Support Factorio 2.0 and Space Age
+
+* Require `setGameConfig` to be called to set the game configuration as there
+  is no longer a single best configuration to use as the default.
+
+* Various bug fixes and enhancements
+
 
 Changes from 0.2 to 0.2.1 (Oct 19, 2024)
 ----------------------------------------